Fix connect app build and relocate some libs
