+++ @connect-shared/lib/attestations/storeUpdatedProjectMetadataAttestation.ts
@@ -1,10 +1,9 @@
 import { DataNotFoundError } from '@charmverse/core/errors';
 import type { OptimismProjectAttestation, OptionalPrismaTransaction } from '@charmverse/core/prisma-client';
 import { prisma } from '@charmverse/core/prisma-client';
-
-import { attestOnchain } from 'lib/credentials/attestOnchain';
-import type { OptimismProjectSnapshotAttestationMetaData } from 'lib/credentials/schemas/optimismProjectSchemas';
-import { getFarcasterProfile } from 'lib/farcaster/getFarcasterProfile';
+import { attestOnchain } from '@root/lib/credentials/attestOnchain';
+import type { OptimismProjectSnapshotAttestationMetaData } from '@root/lib/credentials/schemas/optimismProjectSchemas';
+import { getFarcasterProfile } from '@root/lib/farcaster/getFarcasterProfile';
 
 import { fetchProject } from '../projects/fetchProject';
