+++ apps/connect/components/projects/[id]/ProjectDetailsPage.tsx
@@ -2,7 +2,7 @@ import 'server-only';
 
 import { FarcasterCard } from '@connect/components/common/FarcasterCard';
 import { getCurrentUser } from '@connect/lib/actions/getCurrentUser';
-import type { ConnectProjectDetails } from '@connect/lib/projects/fetchProject';
+import type { ConnectProjectDetails } from '@connect-shared/lib/projects/fetchProject';
 import { Divider, Stack, Typography } from '@mui/material';
 
 import { PageWrapper } from '../../common/PageWrapper';
