+++ apps/connect/lib/actions/createProject.ts
@@ -3,11 +3,11 @@
 import { log } from '@charmverse/core/log';
 import { authActionClient } from '@connect/lib/actions/actionClient';
 import { storeProjectMetadataAndPublishOptimismAttestation } from '@connect-shared/lib/attestations/storeProjectMetadataAndPublishOptimismAttestation';
+import { generateOgImage } from '@connect-shared/lib/projects/generateOgImage';
 import { disableCredentialAutopublish } from '@root/lib/credentials/constants';
 
 import { createConnectProject } from '../projects/createConnectProject';
 import { schema } from '../projects/form';
-import { generateOgImage } from '../projects/generateOgImage';
 
 export const actionCreateProject = authActionClient
   .metadata({ actionName: 'create-project' })
