+++ apps/connect/lib/actions/editProject.ts
@@ -2,15 +2,15 @@
 
 import { log } from '@charmverse/core/log';
 import { authActionClient } from '@connect/lib/actions/actionClient';
+import { storeUpdatedProjectMetadataAttestation } from '@connect-shared/lib/attestations/storeUpdatedProjectMetadataAttestation';
+import { generateOgImage } from '@connect-shared/lib/projects/generateOgImage';
 import { revalidatePath } from 'next/cache';
 
 import { disableCredentialAutopublish } from 'lib/credentials/constants';
 
-import { storeUpdatedProjectMetadataAttestation } from '../attestations/storeUpdatedProjectMetadataAttestation';
 import { editConnectProject } from '../projects/editConnectProject';
 import type { FormValues } from '../projects/form';
 import { schema } from '../projects/form';
-import { generateOgImage } from '../projects/generateOgImage';
 
 export const actionEditProject = authActionClient
   .metadata({ actionName: 'create-project' })
