+++ apps/connect/tsconfig.json
@@ -49,6 +49,6 @@
       "@charmverse/core/errors": ["../../node_modules/@charmverse/core/dist/cjs/errors"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts", "../../@connect-shared/lib/projects/fetchProject.ts", "../../@connect-shared/lib/projects/saveOgImage.ts", "../../@connect-shared/ProjectShareItem.tsx"],
   "exclude": ["../../node_modules", ".next/"]
 }
