+++ apps/sunnyawards/app/p/[path]/edit/page.tsx
@@ -1,8 +1,8 @@
+import { fetchProject } from '@connect-shared/lib/projects/fetchProject';
 import { notFound, redirect } from 'next/navigation';
 
 import { EditProjectPage } from 'components/projects/edit/EditProjectPage';
 import { getCurrentUserAction } from 'lib/profile/getCurrentUserAction';
-import { getProject } from 'lib/projects/getProject';
 
 export default async function EditProject({
   params
@@ -12,7 +12,7 @@ export default async function EditProject({
   };
 }) {
   const [project, user] = await Promise.all([
-    getProject({
+    fetchProject({
       path: params.path
     }),
     getCurrentUserAction()
